Update ABL wall function to work with multiple levels (#229)
Update ABL wall shear stress formulation such that it doesn't require the finest level to cover the entire x/y domain near the ground. Reverts to first cell height sampling for the time being. The ability to sample at any arbitrary height is retained in case it should be resurrected in future.