Add ability to output select components of a field (#236)
Adds a new derived fields variant that can be used to output select
components of a given field.

An example use-case for this would be the CTV Godunov/PLM test which is
essentially a 2D test run in a 3D domain. On GPUs, this causes indeterministic
results on the z-component leading to regression test failures with tight
tolerance. This pull-request also updates that test so that it only outputs
(x, y) components for that test.